requirements-dev.txt
pytest>=7.4
PyYAML>=6.0.1
numpy>=1.26
